feat: refactor assetlinks generate
