+++ lib/src/commands/commands.dart
@@ -1,4 +1,4 @@
-export 'assetlinks_generate_command.dart';
+export 'assetlinks_generate_command/assetlinks_generate_command.dart';
 export 'configurator_generate_command.dart';
 export 'configurator_get_resources_command.dart';
 export 'keystore_commit_command.dart';
